Authorization

+++ app/controllers/categories_controller.rb
@@ -1,11 +1,19 @@
 class CategoriesController < ApplicationController
 
+before_action :require_logged_in
+
   def index
     @categories = Category.all
   end
 
   def new
-    @category = Category.new
+    user = User.find(session[:user_id])
+    if user.admin.eql?(1)
+      @category = Category.new
+    else
+      flash[:error] = "You are not authorized to view this page"
+      render file: 'public/404.html.erb'
+    end
   end
 
   def show
@@ -14,16 +22,26 @@ class CategoriesController < ApplicationController
   end
 
   def edit
-    @category = Category.find(params[:id])
+    user = User.find(session[:user_id])
+    if user.admin.eql?(1)
+      @category = Category.find(params[:id])
+    else
+      flash[:error] = "You are not authorized to view this page"
+      render file: 'public/404.html.erb'
+    end
   end
 
   def update
     @category = Category.find(params[:id])
-    if @category.update(category_params)
+    user = User.find(session[:user_id])
+    if @category.update(category_params) && user.admin.eql?(1)
       flash[:success] = "You updated #{@category.title}"
       redirect_to category_path(@category)
-    else
+    elsif user.admin.eql?(1)
       render :edit
+    else
+      flash[:error] = "You are not authorized to view this page"
+      render file: 'public/404.html.erb'
     end
   end
 
@@ -38,10 +56,16 @@ class CategoriesController < ApplicationController
   end
 
   def destroy
-    category = Category.find(params[:id])
-    category = category.destroy
-    flash[:success] = "#{category.title} was successfully deleted!"
-    redirect_to categories_path
+    user = User.find(session[:user_id])
+    if user.admin.eql?(1)
+      category = Category.find(params[:id])
+      category = category.destroy
+      flash[:success] = "#{category.title} was successfully deleted!"
+      redirect_to categories_path
+    else
+      flash[:error] = "You are not authorized to view this page"
+      render file: 'public/404.html.erb'
+    end
   end
 
   private
